Metadata Tracking Commit
This is the first commit of routines for metadata tracking. It can deduce thinks like motor skill data entry problems, devices which were used to enter data, automated data entry, fraud, unusual data entry behavior, difference from the average regarding data navigation, and finally it can send the data as it is entered rather than on submit. Also an example of a database schema to store the information. It should also be able to record information like user-agent to help fingerprint device