Add ContentData object
As if it wasn't already complex enough, I started passing references to the request in Mono and Flux. At that point I said to myself - Mono and Flux should be lambdas -  self contained anonymous functions. By passing an externel reference it just felt wrong, not to mention asking for a race. So a new object type now exists which should hold a copy of all of the important data from the request. Then that can be passed around and acted on. Down side is that now I need to rewrite the functions to take that object.